backend/requirements.txt
i2cpy
littlefs-python
pywebview
fastapi
uvicorn
pyinstaller